Bump warnings up
Inspired by an oversight Alex Smith caught in another pending PR, increase
warns level somewhat.  Specific -W flags sourced from Yann Collet's blog:
https://fastcompression.blogspot.com/2019/01/compiler-warnings.html

Future work ideas might include incorporating -Wconversion
-Wno-sign-conversion, which requires more change.

Documentation for the GNU make implicit rule (ab)use:
https://www.gnu.org/software/make/manual/html_node/Implicit-Rules.html

+++ tests/t_main.c
@@ -81,7 +81,7 @@ main(int argc, char **argv)
 	hdfs_namenode_delete(nn);
 
 	// Find and run all tests
-	for (int i = 0; i < nelem(suites); i++) {
+	for (size_t i = 0; i < nelem(suites); i++) {
 		Suite *s = suites[i]();
 		SRunner *sr = srunner_create(s);
